pyproject.toml
[project]
name = "photo-org"
version = "0.1.0"
description = "Photo/Video Organizer - Automatically organize media files by date and location"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "geopy>=2.4.1",
    "mutagen>=1.47.0",
    "pillow>=12.0.0",
    "pillow-heif>=1.1.1",
    "pip>=25.3",
    "pyexiftool>=0.5.6",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "tqdm>=4.67.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
]
